src/caption/requirements.txt
# accelerate==1.1.0
einops==0.8.0
elasticsearch==8.15.1
pillow==11.0.0
pytesseract==0.3.13
torch==2.5.1
transformers==4.46.1
timm==1.0.11
